+++ packages/engine/Source/Scene/ArcGISMapService.js
@@ -6,7 +6,7 @@ let defaultTokenCredit;
 const defaultAccessToken =
   "AAPK2f7a9a89c6dc40ad926c3499141fecffWllt95jAiepZyFUkh3j3qU-pKQzn9odSccP_HhOnzggTIrQ90vijnlPCzvS4NJ3t";
 /**
- * Default options for accessing the ArcGIS World Imagery image tile service.
+ * Default options for accessing the ArcGIS image tile service.
  *
  * An ArcGIS access token is required to access ArcGIS image tile layers.
  * A default token is provided for evaluation purposes only.
@@ -14,32 +14,56 @@ const defaultAccessToken =
  * More info can be found in the {@link https://developers.arcgis.com/documentation/mapping-apis-and-services/security/ | ArcGIS developer guide}.
  *
  * @see ArcGisMapServerImageryProvider
- * @namespace ArcGISMapServiceWorldImagery
+ * @namespace ArcGISMapService
  */
-const ArcGISMapServiceWorldImagery = {};
+
+const ArcGISMapService = {};
 /**
  * Gets or sets the default ArcGIS access token.
  *
  * @type {string}
  */
-ArcGISMapServiceWorldImagery.defaultAccessToken = defaultAccessToken;
+ArcGISMapService.defaultAccessToken = defaultAccessToken;
+
 /**
  * Gets or sets the URL of the ArcGIS World Imagery tile service.
  *
  * @type {string|Resource}
  * @default https://ibasemaps-api.arcgis.com/arcgis/rest/services/World_Imagery/MapServer
  */
-ArcGISMapServiceWorldImagery.defaultServer = new Resource({
+ArcGISMapService.defaultWorldImageryServer = new Resource({
   url:
     "https://ibasemaps-api.arcgis.com/arcgis/rest/services/World_Imagery/MapServer",
 });
 
 /**
+ * Gets or sets the URL of the ArcGIS World Hillshade tile service.
  *
- * @param  {string} providedKey
+ * @type {string|Resource}
+ * @default https://ibasemaps-api.arcgis.com/arcgis/rest/services/Elevation/World_Hillshade/MapServer
+ */
+ArcGISMapService.defaultWorldHillshadeServer = new Resource({
+  url:
+    "https://ibasemaps-api.arcgis.com/arcgis/rest/services/Elevation/World_Hillshade/MapServer",
+});
+
+/**
+ * Gets or sets the URL of the ArcGIS World Oceans tile service.
+ *
+ * @type {string|Resource}
+ * @default https://ibasemaps-api.arcgis.com/arcgis/rest/services/Ocean/World_Ocean_Base/MapServer
+ */
+ArcGISMapService.defaultWorldOceanServer = new Resource({
+  url:
+    "https://ibasemaps-api.arcgis.com/arcgis/rest/services/Ocean/World_Ocean_Base/MapServer",
+});
+
+/**
+ *
+ * @param {string} providedKey
  * @return {string|undefined}
  */
-ArcGISMapServiceWorldImagery.getDefaultTokenCredit = function (providedKey) {
+ArcGISMapService.getDefaultTokenCredit = function (providedKey) {
   if (providedKey !== defaultAccessToken) {
     return undefined;
   }
@@ -47,8 +71,8 @@ ArcGISMapServiceWorldImagery.getDefaultTokenCredit = function (providedKey) {
   if (!defined(defaultTokenCredit)) {
     const defaultTokenMessage =
       '<b> \
-            This application is using a default ArcGIS access token. Please assign <i>Cesium.ArcGISMapServiceWorldImagery.defaultAccessToken</i> \
-            with an access token from your ArcGIS account before using the ArcGIS World Imagery tile service. \
+            This application is using a default ArcGIS access token. Please assign <i>Cesium.ArcGISMapService.defaultAccessToken</i> \
+            with an access token from your ArcGIS account before using the ArcGIS tile services. \
             You can sign up for a free ArcGIS Developer account at <a href="https://developers.arcgis.com/">https://developers.arcgis.com/</a>.</b>';
 
     defaultTokenCredit = new Credit(defaultTokenMessage, true);
@@ -56,4 +80,4 @@ ArcGISMapServiceWorldImagery.getDefaultTokenCredit = function (providedKey) {
 
   return defaultTokenCredit;
 };
-export default ArcGISMapServiceWorldImagery;
+export default ArcGISMapService;